refactor(image): image 패키지 분리
