+++ src/main/java/prography/cakeke/server/image/application/port/in/ImageUseCase.java
@@ -1,10 +1,10 @@
-package prography.cakeke.server.store.application.port.in;
+package prography.cakeke.server.image.application.port.in;
 
 import java.util.List;
 
 import org.springframework.web.multipart.MultipartFile;
 
-import prography.cakeke.server.store.adapter.in.web.response.UploadImageResponse;
+import prography.cakeke.server.image.adapter.in.web.response.UploadImageResponse;
 
 public interface ImageUseCase {
     UploadImageResponse postUploadImage(List<MultipartFile> multipartFiles);
