+++ src/main/java/prography/cakeke/server/store/domain/City.java
@@ -1,5 +1,5 @@
 package prography.cakeke.server.store.domain;
 
 public enum City {
-  SEOUL,
+    SEOUL,
 }
